+++ server/src/main/java/com/hobbyt/domain/member/controller/MemberController.java
@@ -21,8 +21,8 @@ import com.hobbyt.domain.member.dto.request.SignupRequest;
 import com.hobbyt.domain.member.dto.request.UpdateMyInfoRequest;
 import com.hobbyt.domain.member.dto.request.UpdatePassword;
 import com.hobbyt.domain.member.dto.response.MyInfoResponse;
-import com.hobbyt.domain.member.service.MemberQueryService;
 import com.hobbyt.domain.member.service.MemberService;
+import com.hobbyt.domain.member.service.query.MemberQueryService;
 import com.hobbyt.global.security.member.MemberDetails;
 
 import lombok.RequiredArgsConstructor;
