+++ server/src/main/java/com/hobbyt/domain/member/service/mail/HtmlTemplate.java
@@ -1,4 +1,4 @@
-package com.hobbyt.domain.member.service;
+package com.hobbyt.domain.member.service.mail;
 
 import java.util.Map;
