+++ server/src/main/java/com/hobbyt/domain/member/service/query/MemberQueryService.java
@@ -1,10 +1,11 @@
-package com.hobbyt.domain.member.service;
+package com.hobbyt.domain.member.service.query;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.hobbyt.domain.member.dto.response.MyInfoResponse;
 import com.hobbyt.domain.member.entity.Member;
+import com.hobbyt.domain.member.service.MemberService;
 
 import lombok.RequiredArgsConstructor;
